split event centralizing and event handling
